# crates/api/assets/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SoraMetrics API",
  "version": "33.0.0",
  "summary": "Read-only analytics of the SORA v2 blockchain.",
  "description": "Public, read-only, no authentication. Data is indexed from the SORA v2 chain (Substrate) and, for balances, holders, pools, staking and governance, read from a node at request time.\n\nPrices are marginal on-chain quotes (`liquidityProxy.quote` of 0.000001 token to DAI, DAI = 1 USD). `illiquid: true` means selling 1 USD worth of the token returns under 0.50 USD: never derive a market cap or a holding value from that price.\n\nA swap is worth its cheaper priced leg (a marginal quote can overvalue the illiquid side by orders of magnitude). Swap rows carry that value in both `in.usd` and `out.usd`; swap volume, trends, top pair and accumulation are sums of it.\n\nToken supply shown by the site is the official SORA circulating figure, not on-chain total issuance.\n\nRate limits are per IP and per route (10–300 requests per minute); a 429 means wait for the next minute. Agents should prefer the MCP endpoint at `/mcp`, which wraps these routes as typed tools.",
  "license": {
   "name": "MIT",
   "identifier": "MIT"
  }
 },
 "servers": [
  {
   "url": "/",
   "description": "This host"
  }
 ],
...
 "paths": {
...
  "/history/extrinsics/{address}": {
   "get": {
    "operationId": "getWalletExtrinsics",
    "summary": "Extrinsics signed by a wallet",
    "description": "Newest first.",
    "tags": [
     "history"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "description": "Paginated envelope.",
         "properties": {
          "data": {
           "type": "array",
           "items": {
            "type": "object",
            "additionalProperties": true
           }
          },
          "total": {
           "type": "integer"
          },
          "page": {
           "type": "integer"
          },
          "totalPages": {
           "type": "integer"
          }
         },
         "required": [
          "data",
          "total",
          "page",
          "totalPages"
         ]
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "address",
      "in": "path",
      "required": true,
      "description": "SS58 address (SORA addresses start with `cn`; other SS58 prefixes are accepted and re-encoded) or 0x-prefixed 32-byte hex.",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "page",
      "in": "query",
      "description": "1-based page. Out-of-range pages are clamped to the last one.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1
      }
     },
     {
      "name": "limit",
      "in": "query",
      "description": "Rows per page.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 100,
       "default": 25
      }
     }
    ]
   }
  },
...
  "/history/global/extrinsics": {
   "get": {
    "operationId": "listExtrinsics",
    "summary": "Latest extrinsics",
    "description": "`timestamp.set` inherents are excluded.",
    "tags": [
     "history"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "description": "Paginated envelope.",
         "properties": {
          "data": {
           "type": "array",
           "items": {
            "type": "object",
            "additionalProperties": true
           }
          },
          "total": {
           "type": "integer"
          },
          "page": {
           "type": "integer"
          },
          "totalPages": {
           "type": "integer"
          }
         },
         "required": [
          "data",
          "total",
          "page",
          "totalPages"
         ]
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "description": "1-based page. Out-of-range pages are clamped to the last one.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1
      }
     },
     {
      "name": "limit",
      "in": "query",
      "description": "Rows per page.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 100,
       "default": 25
      }
     },
     {
      "name": "section",
      "in": "query",
      "description": "Exact pallet name in camelCase (e.g. liquidityProxy).",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "method",
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "block",
      "in": "query",
      "schema": {
       "type": "integer"
      }
     },
     {
      "name": "success",
      "in": "query",
      "schema": {
       "type": "string",
       "enum": [
        "true",
        "false"
       ]
      }
     },
     {
      "name": "timestamp",
      "in": "query",
      "description": "Upper bound, unix milliseconds: only rows at or before it.",
      "schema": {
       "type": "integer"
      }
     }
    ]
   }
  },
...
  "/history/extrinsic/{block}/{index}": {
   "get": {
    "operationId": "getExtrinsic",
    "summary": "Extrinsic detail",
    "description": "Call arguments and emitted events of one extrinsic.",
    "tags": [
     "history"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "block",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1
      }
     },
     {
      "name": "index",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 0
      }
     }
    ]
   }
  },
...
  "/search": {
   "get": {
    "operationId": "search",
    "summary": "Search",
    "description": "Resolves an address, block number, extrinsic id (`block-index`), transaction hash or token symbol.",
    "tags": [
     "chain"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 3,
       "maxLength": 128
      }
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
   "Error": {
    "type": "object",
    "properties": {
     "error": {
      "type": "string"
     },
     "code": {
      "type": "string"
     },
     "message": {
      "type": "string"
     }
    },
    "required": [
     "error"
    ]
   }
  }
 }
}